backend/requirements.txt
pydantic_ai==0.2.18
fastapi==0.115.12
pandas==2.3.0
httpx==0.28.1
tiktoken==0.9.0
psycopg2-binary==2.9.10
pgvector==0.4.1
python-dotenv==1.1.0